Refactor iltm format
