--- tabarena/tabarena/benchmark/models/ag/iltm/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import annotations
