+++ tabarena/tabarena/models/iltm/__init__.py
@@ -1,1 +1,6 @@
 from __future__ import annotations
+
+from tabarena.models.iltm.hpo import gen_iltm
+from tabarena.models.iltm.info import iltm_info, iltm_method_metadata
+
+__all__ = ["gen_iltm", "iltm_info", "iltm_method_metadata"]
